+++ internal/provider/setup_test.go
@@ -1,4 +1,4 @@
-package anxcloud
+package provider
 
 import (
 	"log"